Refactor Entites
