+++ game/src/ecs/components/skill/EnrageSkill.java
@@ -3,7 +3,8 @@ package ecs.components.skill;
 import ecs.components.HealthComponent;
 import ecs.components.VelocityComponent;
 import ecs.entities.Entity;
-import ecs.entities.Hero;
+import ecs.entities.Friendly.Hero;
+
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -56,23 +57,23 @@ public class EnrageSkill extends MagicSkill {
     /**
      * Timer that sets every modified stat back to its original when run out.
      *
-     * @param vCp - VelocityComponent
-     * @param ogX - Original X velocity
-     * @param ogY - Original Y velocity
-     * @param hero - The hero
+     * @param vCp   - VelocityComponent
+     * @param ogX   - Original X velocity
+     * @param ogY   - Original Y velocity
+     * @param hero  - The hero
      * @param ogDmg - Original damage
      */
     public void durationTimer(VelocityComponent vCp, float ogX, float ogY, Hero hero, int ogDmg) {
         Timer timer = new Timer();
         timer.schedule(
-                new TimerTask() {
-                    public void run() {
-                        hero.setDmg(ogDmg);
-                        vCp.setXVelocity(ogX);
-                        vCp.setYVelocity(ogY);
-                        System.out.println("Enrage effect ended.");
-                    }
-                },
-                (long) skillDuration * 1000);
+            new TimerTask() {
+                public void run() {
+                    hero.setDmg(ogDmg);
+                    vCp.setXVelocity(ogX);
+                    vCp.setYVelocity(ogY);
+                    System.out.println("Enrage effect ended.");
+                }
+            },
+            (long) skillDuration * 1000);
     }
 }
