+++ game/src/ecs/entities/Friendly/Hero.java
@@ -1,13 +1,15 @@
-package ecs.entities;
+package ecs.entities.Friendly;
 
 import dslToGame.AnimationBuilder;
 import ecs.components.*;
 import ecs.components.skill.*;
 import ecs.components.xp.ILevelUp;
 import ecs.components.xp.XPComponent;
+import ecs.entities.Entity;
 import ecs.graphic.Animation;
-import java.util.logging.Logger;
 import starter.Game;
+
+import java.util.logging.Logger;
 
 /**
  * The Hero is the player character. It's entity in the ECS. This class helps to setup the hero with
@@ -48,7 +50,9 @@ public class Hero extends Entity implements IOnDeathFunction, ILevelUp {
 
     private static boolean dead = false;
 
-    /** Entity with Components */
+    /**
+     * Entity with Components
+     */
     public Hero() {
         super();
         new PositionComponent(this);
@@ -84,8 +88,8 @@ public class Hero extends Entity implements IOnDeathFunction, ILevelUp {
 
     private void setupFireballSkill() {
         firstSkill =
-                new Skill(
-                        new FireballSkill(SkillTools::getCursorPositionAsPoint), fireballCoolDown);
+            new Skill(
+                new FireballSkill(SkillTools::getCursorPositionAsPoint), fireballCoolDown);
     }
 
     private void setupHealSkill() {
@@ -98,7 +102,9 @@ public class Hero extends Entity implements IOnDeathFunction, ILevelUp {
         sCp.addSkill(thirdSkill);
     }
 
-    /** Modifies the current health by passed amount */
+    /**
+     * Modifies the current health by passed amount
+     */
     public void setHealth(int amount) {
         heroLogger.info("HP before: " + this.hp.getCurrentHealthpoints());
         this.hp.setCurrentHealthpoints(this.hp.getCurrentHealthpoints() + amount);
