+++ game/src/ecs/entities/Monsters/Demon.java
@@ -9,10 +9,11 @@ import ecs.components.ai.transition.FriendlyTransition;
 import ecs.damage.Damage;
 import ecs.damage.DamageType;
 import ecs.entities.Entity;
-import ecs.entities.Monster;
 import ecs.graphic.Animation;
 
-/** The Demon is an enemy monster which inherits from the Monster class. */
+/**
+ * The Demon is an enemy monster which inherits from the Monster class.
+ */
 public class Demon extends Monster {
 
     private boolean fightInRange = false;
