+++ game/src/ecs/entities/Monsters/Imp.java
@@ -9,10 +9,11 @@ import ecs.components.ai.transition.FriendlyTransition;
 import ecs.damage.Damage;
 import ecs.damage.DamageType;
 import ecs.entities.Entity;
-import ecs.entities.Monster;
 import ecs.graphic.Animation;
 
-/** The Imp is an enemy monster which inherits from the Monster class. */
+/**
+ * The Imp is an enemy monster which inherits from the Monster class.
+ */
 public class Imp extends Monster {
 
     private final String pathToIdleLeft = "character/monster/imp/idleLeft";
