+++ game/src/ecs/entities/Traps/BearTrap.java
@@ -1,4 +1,4 @@
-package ecs.entities;
+package ecs.entities.Traps;
 
 import dslToGame.AnimationBuilder;
 import ecs.components.AnimationComponent;
@@ -7,7 +7,9 @@ import ecs.components.HitboxComponent;
 import ecs.components.PositionComponent;
 import ecs.damage.Damage;
 import ecs.damage.DamageType;
+import ecs.entities.Entity;
 import ecs.graphic.Animation;
+
 import java.util.ArrayList;
 
 public class BearTrap extends Trap {
@@ -30,9 +32,9 @@ public class BearTrap extends Trap {
 
     private void setupHitboxComponent() {
         new HitboxComponent(
-                this,
-                (you, other, direction) -> triggerAction(other),
-                (you, other, direction) -> System.out.print(""));
+            this,
+            (you, other, direction) -> triggerAction(other),
+            (you, other, direction) -> System.out.print(""));
     }
 
     void triggerAction(Entity other) {
