+++ game/src/ecs/entities/Traps/Trap.java
@@ -1,6 +1,10 @@
-package ecs.entities;
+package ecs.entities.Traps;
 
-/** Trap Class as a scheme for traps */
+import ecs.entities.Entity;
+
+/**
+ * Trap Class as a scheme for traps
+ */
 public abstract class Trap extends Entity {
     private boolean triggered;
 
@@ -11,7 +15,8 @@ public abstract class Trap extends Entity {
         this.trapDmg = trapDmg;
     }
 
-    public Trap() {}
+    public Trap() {
+    }
 
     public Trap(boolean triggered, boolean repeatable, int trapDmg) {
         this.triggered = triggered;
