+++ game/src/ecs/items/newItems/Bag.java
@@ -3,14 +3,15 @@ package ecs.items.newItems;
 import dslToGame.AnimationBuilder;
 import ecs.components.InventoryComponent;
 import ecs.entities.Entity;
-import ecs.entities.Hero;
+import ecs.entities.Friendly.Hero;
 import ecs.items.IOnCollect;
 import ecs.items.ItemData;
 import ecs.items.ItemType;
 import ecs.items.WorldItemBuilder;
+import starter.Game;
+
 import java.util.ArrayList;
 import java.util.List;
-import starter.Game;
 
 /**
  * A bag is an item that kinda acts like an InventoryComponent, but it only can carry items of the
@@ -24,11 +25,11 @@ public class Bag extends ItemData implements IOnCollect {
 
     public Bag(ItemType itemType) {
         super(
-                ItemType.Bag,
-                AnimationBuilder.buildAnimation("item/world/Bag"),
-                AnimationBuilder.buildAnimation("item/world/Bag"),
-                "Bag",
-                "A bag which is capable of carrying 4 items of the same type.");
+            ItemType.Bag,
+            AnimationBuilder.buildAnimation("item/world/Bag"),
+            AnimationBuilder.buildAnimation("item/world/Bag"),
+            "Bag",
+            "A bag which is capable of carrying 4 items of the same type.");
 
         WorldItemBuilder.buildWorldItem(this);
         inventory = new ArrayList<>(4);
@@ -70,12 +71,16 @@ public class Bag extends ItemData implements IOnCollect {
         return inventory.remove(itemData);
     }
 
-    /** Return the current amount of items in bag */
+    /**
+     * Return the current amount of items in bag
+     */
     public int filledSlots() {
         return inventory.size();
     }
 
-    /** Returns current amount of available slots in bag */
+    /**
+     * Returns current amount of available slots in bag
+     */
     public int emptySlots() {
         return maxSize - inventory.size();
     }
@@ -84,7 +89,9 @@ public class Bag extends ItemData implements IOnCollect {
         return maxSize;
     }
 
-    /** Returns list of items in bag */
+    /**
+     * Returns list of items in bag
+     */
     public List<ItemData> getItems() {
         return new ArrayList<>(inventory);
     }
