+++ game/src/ecs/items/newItems/Greatsword.java
@@ -2,25 +2,27 @@ package ecs.items.newItems;
 
 import dslToGame.AnimationBuilder;
 import ecs.entities.Entity;
-import ecs.entities.Hero;
+import ecs.entities.Friendly.Hero;
 import ecs.items.IOnCollect;
 import ecs.items.ItemData;
 import ecs.items.ItemType;
 import ecs.items.WorldItemBuilder;
 import starter.Game;
 
-/** The greatsword adds 20 damage to the hero if collected. */
+/**
+ * The greatsword adds 20 damage to the hero if collected.
+ */
 public class Greatsword extends ItemData implements IOnCollect {
 
     private final int dmg = 20;
 
     public Greatsword() {
         super(
-                ItemType.Passive,
-                AnimationBuilder.buildAnimation("item/world/Greatsword"),
-                AnimationBuilder.buildAnimation("item/world/Greatsword"),
-                "Greatsword",
-                "Increases the owners damage by 20");
+            ItemType.Passive,
+            AnimationBuilder.buildAnimation("item/world/Greatsword"),
+            AnimationBuilder.buildAnimation("item/world/Greatsword"),
+            "Greatsword",
+            "Increases the owners damage by 20");
 
         WorldItemBuilder.buildWorldItem(this);
         this.setOnCollect(this);
